refactor(tmdb): extract shared cast/crew diffing between movie and episode
UpdateMovieCastAndCrew and UpdateEpisodeCastAndCrew each inlined a
near-identical cast-diff and crew-diff loop, pushing both methods over
SonarCloud's cognitive-complexity threshold (35 and 23 against a limit
of 15).

TMDbLib's ICastCredit/ICrewCredit interfaces (shared by the movie and
episode credit DTOs) make the diffing generic over the credit and
entity types: DiffCast/DiffCrew now hold the add/update/remove logic
once, with per-caller entity construction passed in as a factory. The
episode path's guest-role flag is threaded through DiffCast's optional
applyExtra hook rather than duplicating the loop.

Also drops two redundant `(int)` casts on already-int TMDbLib fields
(episode.EpisodeNumber, reducedEpisode.EpisodeNumber), left over from
the pre-upgrade string-typed API.

Same add/update/remove/save/delete/logging behavior, verified against
the existing test suite (3940/3940 passing).

+++ Shoko.Server/Models/TMDB/TMDB_Episode.cs
@@ -192,7 +192,7 @@ public class TMDB_Episode : TMDB_Base<int>, IEntityMetadata, IEpisode, ITmdbEpis
             UpdateProperty(EnglishTitle, translations is null && !string.IsNullOrEmpty(EnglishTitle) ? EnglishTitle : !string.IsNullOrEmpty(translation?.Data?.Name) ? translation.Data.Name : episode.Name!, v => EnglishTitle = v),
             UpdateProperty(EnglishOverview, !string.IsNullOrEmpty(translation?.Data?.Overview) ? translation.Data.Overview : episode.Overview!, v => EnglishOverview = v),
             UpdateProperty(SeasonNumber, episode.SeasonNumber, v => SeasonNumber = v),
-            UpdateProperty(EpisodeNumber, episode.EpisodeNumber, v => EpisodeNumber = (int)v),
+            UpdateProperty(EpisodeNumber, episode.EpisodeNumber, v => EpisodeNumber = v),
             UpdateProperty(Runtime, episode.Runtime.HasValue ? TimeSpan.FromMinutes(episode.Runtime.Value) : null, v => Runtime = v),
             UpdateProperty(TmdbEpisodeType, string.IsNullOrWhiteSpace(episode.EpisodeType) ? null : episode.EpisodeType.Trim(), v => TmdbEpisodeType = v),
             UpdateProperty(UserRating, episode.VoteAverage, v => UserRating = v),
